resolving conflicts due to:
https://github.com/code-dot-org/code-dot-org/pull/7091
https://github.com/code-dot-org/code-dot-org/pull/7292

Conflicts:
	dashboard/app/controllers/media_proxy_controller.rb

+++ apps/src/applab/commands.js
@@ -12,12 +12,11 @@ var elementUtils = require('./designElements/elementUtils');
 var setPropertyDropdown = require('./setPropertyDropdown');
 var assetPrefix = require('../assetManagement/assetPrefix');
 
-var errorHandler = require('./errorHandler');
-var outputError = errorHandler.outputError;
+var errorHandler = require('../errorHandler');
 var ErrorLevel = errorHandler.ErrorLevel;
 var applabTurtle = require('./applabTurtle');
 var ChangeEventHandler = require('./ChangeEventHandler');
-var colors = require('../sharedJsxStyles').colors;
+var color = require('../color');
 
 var OPTIONAL = true;
 
@@ -40,6 +39,17 @@ var toBeCached = {};
  * @type {EventSandboxer}
  */
 var eventSandboxer = new EventSandboxer();
+
+function outputWarning(errorString) {
+  var line = 1 + window.Applab.JSInterpreter.getNearestUserCodeLine();
+  errorHandler.outputError(errorString, ErrorLevel.WARNING, line);
+}
+
+function outputError(errorString) {
+  var line = 1 + window.Applab.JSInterpreter.getNearestUserCodeLine();
+  errorHandler.outputError(errorString, ErrorLevel.ERROR, line);
+}
+
 
 /**
  * @param value
@@ -95,16 +105,20 @@ function apiValidateType(opts, funcName, varName, varValue, expectedType, opt) {
       case 'array':
         properType = Array.isArray(varValue);
         break;
+      case 'record':
+        // Validate that we have a data record. These must be objects, and
+        // not arrays
+        properType = typeof varValue === 'object' && !Array.isArray(varValue);
+        break;
       default:
         properType = (typeof varValue === expectedType);
         break;
     }
     properType = properType || (opt === OPTIONAL && (typeof varValue === 'undefined'));
     if (!properType) {
-      var line = 1 + Applab.JSInterpreter.getNearestUserCodeLine();
-      var errorString = funcName + "() " + varName + " parameter value (" +
-        varValue + ") is not a " + expectedType + ".";
-      outputError(errorString, ErrorLevel.WARNING, line);
+      outputWarning(funcName + "() " + varName + " parameter value (" +
+        varValue + ") is not a " + expectedType + ".");
+
     }
     opts[validatedTypeKey] = properType;
   }
@@ -123,10 +137,8 @@ function apiValidateTypeAndRange(opts, funcName, varName, varValue,
     }
     inRange = inRange || (opt === OPTIONAL && (typeof varValue === 'undefined'));
     if (!inRange) {
-      var line = 1 + Applab.JSInterpreter.getNearestUserCodeLine();
-      var errorString = funcName + "() " + varName + " parameter value (" +
-        varValue + ") is not in the expected range.";
-      outputError(errorString, ErrorLevel.WARNING, line);
+      outputWarning(funcName + "() " + varName + " parameter value (" +
+        varValue + ") is not in the expected range.");
     }
     opts[validatedRangeKey] = inRange;
   }
@@ -137,10 +149,8 @@ function apiValidateActiveCanvas(opts, funcName) {
   if (!opts || typeof opts[validatedActiveCanvasKey] === 'undefined') {
     var activeCanvas = Boolean(Applab.activeCanvas);
     if (!activeCanvas) {
-      var line = 1 + Applab.JSInterpreter.getNearestUserCodeLine();
-      var errorString = funcName + "() called without an active canvas. Call " +
-        "createCanvas() first.";
-      outputError(errorString, ErrorLevel.WARNING, line);
+      outputWarning(funcName + "() called without an active canvas. Call " +
+        "createCanvas() first.");
     }
     if (opts) {
       opts[validatedActiveCanvasKey] = activeCanvas;
@@ -168,17 +178,15 @@ function apiValidateDomIdExistence(opts, funcName, varName, id, shouldExist) {
     var valid = !existsOutsideApplab && (shouldExist == existsInApplab);
 
     if (!valid) {
-      var line = 1 + Applab.JSInterpreter.getNearestUserCodeLine();
       var errorString = "";
       if (existsOutsideApplab) {
         errorString = funcName + "() " + varName + " parameter refers to an id (" + id +
             ") which is already in use outside of Applab. Choose a different id.";
         throw new Error(errorString);
       } else {
-        errorString = funcName + "() " + varName +
+        outputWarning(funcName + "() " + varName +
             " parameter refers to an id (" + id + ") which " +
-            (existsInApplab ? "already exists." : "does not exist.");
-        outputError(errorString, ErrorLevel.WARNING, line);
+            (existsInApplab ? "already exists." : "does not exist."));
       }
     }
     opts[validatedDomKey] = valid;
@@ -191,7 +199,7 @@ applabCommands.setScreen = function (opts) {
   apiValidateDomIdExistence(opts, 'setScreen', 'screenId', opts.screenId, true);
   var element = document.getElementById(opts.screenId);
   var divApplab = document.getElementById('divApplab');
-  if (!divApplab.contains(element)) {
+  if (!element || (element.parentNode != divApplab)) {
     return;
   }
 
@@ -199,13 +207,12 @@ applabCommands.setScreen = function (opts) {
 };
 
 function reportUnsafeHtml(removed, unsafe, safe, warnings) {
-  var currentLineNumber = getCurrentLineNumber(Applab.JSInterpreter);
   var msg = "The following lines of HTML were modified or removed:\n" + removed +
       "\noriginal html:\n" + unsafe + "\nmodified html:\n" + safe;
   if (warnings.length > 0) {
     msg += '\nwarnings:\n' + warnings.join('\n');
   }
-  outputError(msg, ErrorLevel.WARNING, currentLineNumber);
+  outputWarning(msg);
 }
 
 applabCommands.container = function (opts) {
@@ -237,8 +244,8 @@ applabCommands.button = function (opts) {
   var textNode = document.createTextNode(opts.text);
   newButton.id = opts.elementId;
   newButton.style.position = 'relative';
-  newButton.style.color = colors.white;
-  newButton.style.backgroundColor = colors.teal;
+  newButton.style.color = color.white;
+  newButton.style.backgroundColor = color.applab_button_teal;
 
   return Boolean(newButton.appendChild(textNode) &&
     Applab.activeScreen().appendChild(newButton));
@@ -870,8 +877,8 @@ applabCommands.dropdown = function (opts) {
   }
   newSelect.id = opts.elementId;
   newSelect.style.position = 'relative';
-  newSelect.style.color = colors.white;
-  newSelect.style.backgroundColor = colors.teal;
+  newSelect.style.color = color.white;
+  newSelect.style.backgroundColor = color.applab_button_teal;
 
   return Boolean(Applab.activeScreen().appendChild(newSelect));
 };
@@ -880,7 +887,7 @@ applabCommands.getAttribute = function (opts) {
   var divApplab = document.getElementById('divApplab');
   var element = document.getElementById(opts.elementId);
   var attribute = String(opts.attribute);
-  return divApplab.contains(element) ? element[attribute] : false;
+  return divApplab.contains(element) ? String(element[attribute]) : false;
 };
 
 // Whitelist of HTML Element attributes which can be modified, to
@@ -1195,9 +1202,7 @@ applabCommands.setProperty = function(opts) {
 
   var info = setPropertyDropdown.getInternalPropertyInfo(element, property);
   if (!info) {
-    var currentLineNumber = getCurrentLineNumber(Applab.JSInterpreter);
-    outputError('Cannot set property "' + property + '" on element "' + elementId + '".',
-      ErrorLevel.ERROR, currentLineNumber);
+    outputError('Cannot set property "' + property + '" on element "' + elementId + '".');
     return;
   }
 
@@ -1430,10 +1435,20 @@ applabCommands.createRecord = function (opts) {
   // PARAMNAME: createRecord: table vs. tableName
   // PARAMNAME: createRecord: callback vs. callbackFunction
   apiValidateType(opts, 'createRecord', 'table', opts.table, 'string');
-  apiValidateType(opts, 'createRecord', 'record', opts.record, 'object');
-  apiValidateType(opts, 'createRecord', 'record.id', opts.record.id, 'undefined');
+  var validRecord = apiValidateType(opts, 'createRecord', 'record', opts.record, 'record');
   apiValidateType(opts, 'createRecord', 'callback', opts.onSuccess, 'function', OPTIONAL);
   apiValidateType(opts, 'createRecord', 'onError', opts.onError, 'function', OPTIONAL);
+  if (!validRecord) {
+    return;
+  }
+  if (!opts.table) {
+    outputError('missing required parameter "tableName"');
+    return;
+  }
+  if (opts.record.id) {
+    outputError('record must not have an "id" property');
+    return;
+  }
   var onSuccess = applabCommands.handleCreateRecord.bind(this, opts);
   var onError = errorHandler.handleError.bind(this, opts);
   AppStorage.createRecord(opts.table, opts.record, onSuccess, onError);
@@ -1522,6 +1537,10 @@ applabCommands.readRecords = function (opts) {
   apiValidateType(opts, 'readRecords', 'searchTerms', opts.searchParams, 'object');
   apiValidateType(opts, 'readRecords', 'callback', opts.onSuccess, 'function');
   apiValidateType(opts, 'readRecords', 'onError', opts.onError, 'function', OPTIONAL);
+  if (!opts.table) {
+    outputError('missing required parameter "tableName"');
+    return;
+  }
   var onSuccess = applabCommands.handleReadRecords.bind(this, opts);
   var onError = errorHandler.handleError.bind(this, opts);
   AppStorage.readRecords(opts.table, opts.searchParams, onSuccess, onError);
@@ -1537,10 +1556,21 @@ applabCommands.updateRecord = function (opts) {
   // PARAMNAME: updateRecord: table vs. tableName
   // PARAMNAME: updateRecord: callback vs. callbackFunction
   apiValidateType(opts, 'updateRecord', 'table', opts.table, 'string');
-  apiValidateType(opts, 'updateRecord', 'record', opts.record, 'object');
+  var validRecord = apiValidateType(opts, 'updateRecord', 'record', opts.record, 'record');
   apiValidateTypeAndRange(opts, 'updateRecord', 'record.id', opts.record.id, 'number', 1, Infinity);
   apiValidateType(opts, 'updateRecord', 'callback', opts.onComplete, 'function', OPTIONAL);
   apiValidateType(opts, 'updateRecord', 'onError', opts.onError, 'function', OPTIONAL);
+  if (!validRecord) {
+    return;
+  }
+  if (!opts.table) {
+    outputError('missing required parameter "tableName"');
+    return;
+  }
+  if (opts.record.id === undefined) {
+    outputError('missing required property "id"');
+    return;
+  }
   var onComplete = applabCommands.handleUpdateRecord.bind(this, opts);
   var onError = errorHandler.handleError.bind(this, opts);
   AppStorage.updateRecord(opts.table, opts.record, onComplete, onError);
@@ -1556,10 +1586,21 @@ applabCommands.deleteRecord = function (opts) {
   // PARAMNAME: deleteRecord: table vs. tableName
   // PARAMNAME: deleteRecord: callback vs. callbackFunction
   apiValidateType(opts, 'deleteRecord', 'table', opts.table, 'string');
-  apiValidateType(opts, 'deleteRecord', 'record', opts.record, 'object');
+  var validRecord = apiValidateType(opts, 'deleteRecord', 'record', opts.record, 'record');
   apiValidateTypeAndRange(opts, 'deleteRecord', 'record.id', opts.record.id, 'number', 1, Infinity);
   apiValidateType(opts, 'deleteRecord', 'callback', opts.onComplete, 'function', OPTIONAL);
   apiValidateType(opts, 'deleteRecord', 'onError', opts.onError, 'function', OPTIONAL);
+  if (!validRecord) {
+    return;
+  }
+  if (!opts.table) {
+    outputError('missing required parameter "tableName"');
+    return;
+  }
+  if (opts.record.id === undefined) {
+    outputError('missing required property "id"');
+    return;
+  }
   var onComplete = applabCommands.handleDeleteRecord.bind(this, opts);
   var onError = errorHandler.handleError.bind(this, opts);
   AppStorage.deleteRecord(opts.table, opts.record, onComplete, onError);
@@ -1569,6 +1610,13 @@ applabCommands.handleDeleteRecord = function(opts, success) {
   if (opts.onComplete) {
     opts.onComplete.call(null, success);
   }
+};
+
+applabCommands.onRecordEvent = function (opts) {
+  apiValidateType(opts, 'onRecordEvent', 'table', opts.table, 'string');
+  apiValidateType(opts, 'onRecordEvent', 'callback', opts.onRecord, 'function');
+  var onError = errorHandler.handleError.bind(this, opts);
+  AppStorage.onRecordEvent(opts.table, opts.onRecord, onError);
 };
 
 applabCommands.getUserId = function (opts) {
@@ -1601,12 +1649,6 @@ applabCommands.drawChart = function (opts) {
   apiValidateType(opts, 'drawChart', 'callback', opts.callback, 'function', OPTIONAL);
   apiValidateDomIdExistence(opts, 'drawChart', 'chartId', opts.chartId, true);
 
-  // Bind a reference to the current interpreter so we can later make sure we're
-  // not doing anything asynchronous across re-runs.
-  var jsInterpreter = Applab.JSInterpreter;
-
-  var currentLineNumber = getCurrentLineNumber(jsInterpreter);
-
   var chartApi = new ChartApi();
 
   /**
@@ -1621,7 +1663,7 @@ applabCommands.drawChart = function (opts) {
   var onSuccess = function () {
     stopLoadingSpinnerFor(opts.chartId);
     chartApi.warnings.forEach(function (warning) {
-      outputError(warning.message, ErrorLevel.WARNING, currentLineNumber);
+      outputWarning(warning.message);
     });
     if (opts.callback) {
       opts.callback.call(null);
@@ -1636,7 +1678,7 @@ applabCommands.drawChart = function (opts) {
    */
   var onError = function (error) {
     stopLoadingSpinnerFor(opts.chartId);
-    outputError(error.message, ErrorLevel.ERROR, currentLineNumber);
+    outputError(error.message);
   };
 
   startLoadingSpinnerFor(opts.chartId);
@@ -1673,12 +1715,6 @@ applabCommands.drawChartFromRecords = function (opts) {
   apiValidateType(opts, 'drawChartFromRecords', 'callback', opts.callback, 'function', OPTIONAL);
   apiValidateDomIdExistence(opts, 'drawChartFromRecords', 'chartId', opts.chartId, true);
 
-  // Bind a reference to the current interpreter so we can later make sure we're
-  // not doing anything asynchronous across re-runs.
-  var jsInterpreter = Applab.JSInterpreter;
-
-  var currentLineNumber = getCurrentLineNumber(jsInterpreter);
-
   var chartApi = new ChartApi();
 
   /**
@@ -1693,7 +1729,7 @@ applabCommands.drawChartFromRecords = function (opts) {
   var onSuccess = function () {
     stopLoadingSpinnerFor(opts.chartId);
     chartApi.warnings.forEach(function (warning) {
-      outputError(warning.message, ErrorLevel.WARNING, currentLineNumber);
+      outputWarning(warning.message);
     });
     if (opts.callback) {
       opts.callback.call(null);
@@ -1708,7 +1744,7 @@ applabCommands.drawChartFromRecords = function (opts) {
    */
   var onError = function (error) {
     stopLoadingSpinnerFor(opts.chartId);
-    outputError(error.message, ErrorLevel.ERROR, currentLineNumber);
+    outputError(error.message);
   };
 
   startLoadingSpinnerFor(opts.chartId);
@@ -1752,13 +1788,3 @@ function stopLoadingSpinnerFor(elementId) {
     return !(/loading/i.test(x));
   }).join(' ');
 }
-
-/**
- * For the provided interpreter, get the nearest line number in user code
- * up the stack from the last executed command.
- * @param {JSInterpreter} jsInterpreter
- * @returns {number}
- */
-var getCurrentLineNumber = function (jsInterpreter) {
-  return 1 + jsInterpreter.getNearestUserCodeLine();
-};
